+++ ecd-20200930/alibabacloud_ecd20200930/models/_modify_qos_entries_request.py
@@ -10,16 +10,20 @@ class ModifyQosEntriesRequest(DaraModel):
     def __init__(
         self,
         auth_android_id: List[str] = None,
+        auth_desktop_group_id: List[str] = None,
         auth_desktop_id: List[str] = None,
         qos_rule_id: str = None,
         revoke_android_id: List[str] = None,
+        revoke_desktop_group_id: List[str] = None,
         revoke_desktop_id: List[str] = None,
     ):
         self.auth_android_id = auth_android_id
+        self.auth_desktop_group_id = auth_desktop_group_id
         self.auth_desktop_id = auth_desktop_id
         # This parameter is required.
         self.qos_rule_id = qos_rule_id
         self.revoke_android_id = revoke_android_id
+        self.revoke_desktop_group_id = revoke_desktop_group_id
         self.revoke_desktop_id = revoke_desktop_id
 
     def validate(self):
@@ -33,6 +37,9 @@ class ModifyQosEntriesRequest(DaraModel):
         if self.auth_android_id is not None:
             result['AuthAndroidId'] = self.auth_android_id
 
+        if self.auth_desktop_group_id is not None:
+            result['AuthDesktopGroupId'] = self.auth_desktop_group_id
+
         if self.auth_desktop_id is not None:
             result['AuthDesktopId'] = self.auth_desktop_id
 
@@ -41,6 +48,9 @@ class ModifyQosEntriesRequest(DaraModel):
 
         if self.revoke_android_id is not None:
             result['RevokeAndroidId'] = self.revoke_android_id
+
+        if self.revoke_desktop_group_id is not None:
+            result['RevokeDesktopGroupId'] = self.revoke_desktop_group_id
 
         if self.revoke_desktop_id is not None:
             result['RevokeDesktopId'] = self.revoke_desktop_id
@@ -52,6 +62,9 @@ class ModifyQosEntriesRequest(DaraModel):
         if m.get('AuthAndroidId') is not None:
             self.auth_android_id = m.get('AuthAndroidId')
 
+        if m.get('AuthDesktopGroupId') is not None:
+            self.auth_desktop_group_id = m.get('AuthDesktopGroupId')
+
         if m.get('AuthDesktopId') is not None:
             self.auth_desktop_id = m.get('AuthDesktopId')
 
@@ -60,6 +73,9 @@ class ModifyQosEntriesRequest(DaraModel):
 
         if m.get('RevokeAndroidId') is not None:
             self.revoke_android_id = m.get('RevokeAndroidId')
+
+        if m.get('RevokeDesktopGroupId') is not None:
+            self.revoke_desktop_group_id = m.get('RevokeDesktopGroupId')
 
         if m.get('RevokeDesktopId') is not None:
             self.revoke_desktop_id = m.get('RevokeDesktopId')